pyproject.toml
[project]
name = "malcolm"
version = "0.1.0"
description = "Transparent LLM API monitoring proxy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20",
]

[project.scripts]
malcolm = "malcolm.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/malcolm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
